Authorization filter and http status response added in all endpoints.

+++ src/main/java/com/utn/TPFinal/controllers/CallController.java
@@ -1,73 +1,114 @@
 package com.utn.TPFinal.controllers;
 
+import com.utn.TPFinal.model.Enum.UserTypeEnum;
 import com.utn.TPFinal.model.entities.*;
 import com.utn.TPFinal.services.CallService;
+import com.utn.TPFinal.session.SessionManager;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
 @RestController("")
-@RequestMapping("/Call")
+@RequestMapping("/api/Call")
 public class CallController {
 
     private final CallService callService;
+    private final SessionManager sessionManager;
 
     @Autowired
-    public CallController(CallService callService) {
+    public CallController(CallService callService, SessionManager sessionManager) {
         this.callService = callService;
+        this.sessionManager = sessionManager;
     }
 
 
     @GetMapping("/")
-    public List<Call> getAll(){
+    public ResponseEntity<List<Call>> getAll(@RequestHeader("Authorization") String sessionToken){
         try{
-            return callService.getAll();
+            User user = sessionManager.getCurrentUser(sessionToken);
+
+            if(user!=null && user.getUserType().getName().toUpperCase().equals(UserTypeEnum.EMPLEADO.name())){
+                List<Call> calls = callService.getAll();
+                return (calls.size() > 0) ? ResponseEntity.ok(calls) : ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+            }
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
         }catch (Exception ex){
             throw ex;
         }
     }
 
     @GetMapping("/{id}")
-    public Call getById(@PathVariable Integer id){
+    public ResponseEntity<Call> getById(@RequestHeader("Authorization") String sessionToken,@PathVariable Integer id){
         try{
-            return callService.getById(id);
+            User user = sessionManager.getCurrentUser(sessionToken);
+
+            if(user!=null && user.getUserType().getName().toUpperCase().equals(UserTypeEnum.EMPLEADO.name())){
+                Call call= callService.getById(id);
+                return (call != null) ? ResponseEntity.ok(call) : ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+            }
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
         }catch (Exception ex){
             throw ex;
         }
     }
 
     @GetMapping("/User/{id}")
-    public List<Call> getByUserId(@PathVariable Integer id){
+    public ResponseEntity<List<Call>> getByUserId(@RequestHeader("Authorization") String sessionToken,@PathVariable Integer id){
         try{
-            return callService.getByUserId(id);
+            User user = sessionManager.getCurrentUser(sessionToken);
+
+            if(user!=null && (user.getUserType().getName().toUpperCase().equals(UserTypeEnum.EMPLEADO.name()) || user.getId() == id)){
+                List<Call> calls = callService.getByUserId(id);
+                return (calls.size() > 0) ? ResponseEntity.ok(calls) : ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+            }
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
         }catch (Exception ex){
             throw ex;
         }
     }
 
     @PostMapping("/")
-    public Integer add(@RequestBody Call call){
+    public ResponseEntity<Integer> add(@RequestHeader("Authorization") String sessionToken,@RequestBody Call call){
         try{
-           return callService.add(call);
+            User user = sessionManager.getCurrentUser(sessionToken);
+
+            if(user!=null && user.getUserType().getName().toUpperCase().equals(UserTypeEnum.INFRAESTRUCTURA.name())){
+                return ResponseEntity.status(HttpStatus.CREATED).body(callService.add(call));
+            }
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
         }catch (Exception ex){
             throw ex;
         }
     }
 
     @PutMapping("/")
-    public void update(@RequestBody Call call) throws Exception {
+    public ResponseEntity update(@RequestHeader("Authorization") String sessionToken,@RequestBody Call call) throws Exception {
         try{
-            callService.update(call);
+            User user = sessionManager.getCurrentUser(sessionToken);
+
+            if(user!=null && user.getUserType().getName().toUpperCase().equals(UserTypeEnum.INFRAESTRUCTURA.name())){
+                callService.update(call);
+                return ResponseEntity.ok().build();
+            }
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
         }catch (Exception ex){
             throw ex;
         }
     }
 
     @DeleteMapping("/{id}")
-    public void remove(@PathVariable Integer id){
+    public ResponseEntity remove(@RequestHeader("Authorization") String sessionToken,@PathVariable Integer id){
         try{
-            callService.remove(id);
+            User user = sessionManager.getCurrentUser(sessionToken);
+
+            if(user!=null && user.getUserType().getName().toUpperCase().equals(UserTypeEnum.INFRAESTRUCTURA.name())){
+                callService.remove(id);
+                return ResponseEntity.ok().build();
+            }
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
         }catch (Exception ex){
             throw ex;
         }
