+++ src/main/java/com/utn/TPFinal/controllers/PhoneLineTypeController.java
@@ -1,63 +1,94 @@
 package com.utn.TPFinal.controllers;
 
+import com.utn.TPFinal.model.Enum.UserTypeEnum;
 import com.utn.TPFinal.model.entities.PhoneLineType;
+import com.utn.TPFinal.model.entities.User;
 import com.utn.TPFinal.services.PhoneLineTypeService;
+import com.utn.TPFinal.session.SessionManager;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import java.util.List;
 
 @RestController("")
-@RequestMapping("/PhoneLineType")
+@RequestMapping("/api/PhoneLineType")
 public class PhoneLineTypeController {
 
     private final PhoneLineTypeService phoneLineTypeService;
+    private final SessionManager sessionManager;
 
     @Autowired
-    public PhoneLineTypeController(PhoneLineTypeService phoneLineTypeService) {
+    public PhoneLineTypeController(PhoneLineTypeService phoneLineTypeService, SessionManager sessionManager) {
         this.phoneLineTypeService = phoneLineTypeService;
+        this.sessionManager = sessionManager;
     }
 
 
     @GetMapping("/")
-    public List<PhoneLineType> getAll(){
+    public ResponseEntity<List<PhoneLineType>> getAll(@RequestHeader("Authorization") String sessionToken){
         try{
-            return phoneLineTypeService.getAll();
+            List<PhoneLineType> phoneLineTypes = phoneLineTypeService.getAll();
+            return (phoneLineTypes.size() > 0) ? ResponseEntity.ok(phoneLineTypes) : ResponseEntity.status(HttpStatus.NO_CONTENT).build();
         }catch (Exception ex){
             throw ex;
         }
     }
 
     @GetMapping("/{id}")
-    public PhoneLineType getById(@PathVariable Integer id){
+    public ResponseEntity<PhoneLineType> getById(@RequestHeader("Authorization") String sessionToken,@PathVariable Integer id){
         try{
-            return phoneLineTypeService.getById(id);
+            User user = sessionManager.getCurrentUser(sessionToken);
+
+            if(user!=null && user.getUserType().getName().toUpperCase().equals(UserTypeEnum.EMPLEADO.name())){
+                PhoneLineType phoneLineType = phoneLineTypeService.getById(id);
+                return (phoneLineType != null) ? ResponseEntity.ok(phoneLineType) : ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+            }
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
         }catch (Exception ex){
             throw ex;
         }
     }
 
     @PostMapping("/")
-    public Integer add(@RequestBody PhoneLineType phoneLineType){
+    public ResponseEntity<Integer> add(@RequestHeader("Authorization") String sessionToken,@RequestBody PhoneLineType phoneLineType){
         try{
-            return phoneLineTypeService.add(phoneLineType);
+            User user = sessionManager.getCurrentUser(sessionToken);
+
+            if(user!=null && user.getUserType().getName().toUpperCase().equals(UserTypeEnum.EMPLEADO.name())){
+                return ResponseEntity.status(HttpStatus.CREATED).body(phoneLineTypeService.add(phoneLineType));
+            }
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
         }catch (Exception ex){
             throw ex;
         }
     }
 
     @PutMapping("/")
-    public void update(@RequestBody PhoneLineType phoneLineType) throws Exception {
+    public ResponseEntity update(@RequestHeader("Authorization") String sessionToken,@RequestBody PhoneLineType phoneLineType) throws Exception {
         try{
-            phoneLineTypeService.update(phoneLineType);
+            User user = sessionManager.getCurrentUser(sessionToken);
+
+            if(user!=null && user.getUserType().getName().toUpperCase().equals(UserTypeEnum.EMPLEADO.name())){
+                phoneLineTypeService.update(phoneLineType);
+                return ResponseEntity.ok().build();
+            }
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
         }catch (Exception ex){
             throw ex;
         }
     }
 
     @DeleteMapping("/{id}")
-    public void remove(@PathVariable Integer id){
+    public ResponseEntity remove(@RequestHeader("Authorization") String sessionToken,@PathVariable Integer id){
         try{
-            phoneLineTypeService.remove(id);
+            User user = sessionManager.getCurrentUser(sessionToken);
+
+            if(user!=null && user.getUserType().getName().toUpperCase().equals(UserTypeEnum.EMPLEADO.name())){
+                phoneLineTypeService.remove(id);
+                return ResponseEntity.ok().build();
+            }
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
         }catch (Exception ex){
             throw ex;
         }
